refactor: move steam related things to it's own module.
